move all eval files to sub dir
