+++ ymir/command/mir/tools/eval/ins_seg_eval_coco.py
@@ -1,0 +1,7 @@
+
+from mir.protos import mir_command_pb2 as mirpb
+
+
+def evaluate(prediction: mirpb.SingleTaskAnnotations, ground_truth: mirpb.SingleTaskAnnotations,
+             config: mirpb.EvaluateConfig, assets_metadata: mirpb.MirMetadatas) -> mirpb.Evaluation:
+    pass
