+++ ymir/command/tests/unit/test_tools_seg_eval.py
@@ -8,7 +8,7 @@ from google.protobuf import json_format
 import numpy as np
 
 from mir.protos import mir_command_pb2 as mirpb
-from mir.tools import det_eval_ops
+from mir.tools.eval import det_eval_ops
 from tests import utils as test_utils
